refactor(app): change hook to util
