+++ apps/app/src/utils/pageScroll.ts
@@ -1,0 +1,11 @@
+const classes = ["overflow-hidden", "sm:overflow-scroll"];
+
+export const lockScroll = () => {
+	document.body.style.paddingRight = `${window.innerWidth - document.body.offsetWidth}px`;
+	document.body.classList.add(...classes);
+};
+
+export const unlockScroll = () => {
+	document.body.style.paddingRight = "";
+	document.body.classList.remove(...classes);
+};
